Fix for PDI breaking changes
- xyz.openbmc_project.Inventory.Item.Dimm MemorySizeInKB type was
  changed from uint32 to size.
- xyz.openbmc_project.Common.UUID UUID field was renamed (in the
  generated server binding code) from uUID to uuid. Use the sdbus++
  preprocessor define to tell if this name change is in effect or not,
  in order to build with multiple versions of sdbusplus.

Tested: Build passes.

+++ include/dimm.hpp
@@ -66,7 +66,7 @@ class Dimm :
     void memoryInfoUpdate(void);
 
     uint16_t memoryDataWidth(uint16_t value) override;
-    uint32_t memorySizeInKB(uint32_t value) override;
+    size_t memorySizeInKB(size_t value) override;
     std::string memoryDeviceLocator(std::string value) override;
     DeviceType memoryType(DeviceType value) override;
     std::string memoryTypeDetail(std::string value) override;
